Warn on dangerously low voltage
Capacity monitoring on Lenovo dual-battery systems doesn't work reliably
when one of the cells isn't discharged or topped off for long periods of
time, i.e. when the charge thresholds are set to 70% / 80% and the
charge controller is only ever discharging and charging the other
battery. Eventually the voltage will drop below the safe threshold, even
though it'll claim that the cell still has a reasonable amount of energy
left and the charge controller won't top it off. I'm guessing that the
charge controller firmware is largely to blame, but this can probably be
mitigated somewhat in software. For now, just indicate that something is
wrong.